requirements.txt
Django==1.11.7
djangorestframework==3.7.3
ipython==6.2.1
Pillow